% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\docType{data}
\name{ImportMode}
\alias{ImportMode}
\title{ImportMode}
\format{
An object of class \code{list} of length 2.
}
\usage{
ImportMode
}
\description{
The following import modes are supported:
\itemize{
  \item{\code{OVERWRITE}: Overwrite any existing data.}
  \item{\code{MERGE}: Merge the new data with the existing data based on resource ID.}
}
}
\keyword{datasets}
