% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LOINC_URI}
\alias{LOINC_URI}
\title{LOINC system URI}
\usage{
LOINC_URI
}
\description{
The URI of the LOINC code system: \code{http://loinc.org}.
}
\seealso{
\href{https://terminology.hl7.org/LOINC.html}{Using LOINC with HL7 Standards}
}
