% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminology_import.R
\name{pathling_import_fhir_terminology}
\alias{pathling_import_fhir_terminology}
\title{Import FHIR terminology resources}
\usage{
pathling_import_fhir_terminology(pc, source, storage_path)
}
\arguments{
\item{pc}{The PathlingContext object.}

\item{source}{The path to a JSON file, a directory of JSON files, or a FHIR NPM package, on any
filesystem accessible through the Hadoop FileSystem API.}

\item{storage_path}{The terminology store location, created if absent.}
}
\value{
The PathlingContext object, invisibly.
}
\description{
Imports FHIR R4 CodeSystem, ValueSet, and ConceptMap resources into a local terminology store.
The source may be a JSON file, a directory of JSON files, or a FHIR NPM package (.tgz).
}
\seealso{
Other terminology import functions:
\code{\link[=pathling_import_snomed]{pathling_import_snomed()}}
}
\concept{terminology import functions}
