% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{pathling_install_spark}
\alias{pathling_install_spark}
\title{Install Spark}
\usage{
pathling_install_spark()
}
\value{
List with information about the installed version.
}
\description{
Installs the version of Spark/Hadoop defined in the package metadata using the
\code{sparklyr::spark_install} function.
}
\seealso{
Other installation functions:
\code{\link[=pathling_is_spark_installed]{pathling_is_spark_installed()}},
\code{\link[=pathling_spark_info]{pathling_spark_info()}},
\code{\link[=pathling_version]{pathling_version()}}
}
\concept{installation functions}
