% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{pathling_spark}
\alias{pathling_spark}
\title{Get the Spark session}
\usage{
pathling_spark(pc)
}
\arguments{
\item{pc}{The PathlingContext object.}
}
\value{
The Spark connection associated with this Pathling context.
}
\description{
Returns the Spark connection associated with a Pathling context.
}
\seealso{
Other context lifecycle functions:
\code{\link[=pathling_connect]{pathling_connect()}},
\code{\link[=pathling_disconnect]{pathling_disconnect()}},
\code{\link[=pathling_disconnect_all]{pathling_disconnect_all()}}
}
\concept{context lifecycle functions}
