% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udfs.R
\name{PropertyType}
\alias{PropertyType}
\title{Coding property data types}
\usage{
PropertyType
}
\description{
The following data types are supported:
\itemize{
  \item \code{STRING} - A string value.
  \item \code{INTEGER} - An integer value.
  \item \code{BOOLEAN} - A boolean value.
  \item \code{DECIMAL} - A decimal value.
  \item \code{DATETIME} - A date/time value.
  \item \code{CODE} - A code value.
  \item \code{CODING} - A Coding value.
}
}
\seealso{
\href{https://hl7.org/fhir/R4/datatypes.html}{FHIR R4 - Data Types}
}
