% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{SaveMode}
\alias{SaveMode}
\title{SaveMode}
\usage{
SaveMode
}
\description{
The following save modes are supported:
\itemize{
  \item{\code{OVERWRITE}: Overwrite any existing data.}
  \item{\code{APPEND}: Append the new data to the existing data.}
  \item{\code{IGNORE}: Only save the data if the file does not already exist.}
  \item{\code{ERROR}: Raise an error if the file already exists.}
  \item{\code{MERGE}: Merge the new data with the existing data based on resource ID.}
}
}
