% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SNOMED_URI}
\alias{SNOMED_URI}
\title{SNOMED CT system URI}
\usage{
SNOMED_URI
}
\description{
The URI of the SNOMED CT code system: \code{http://snomed.info/sct}.
}
\seealso{
\href{https://terminology.hl7.org/SNOMEDCT.html}{Using SNOMED CT with HL7 Standards}
}
