% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{StorageType}
\alias{StorageType}
\title{Terminology cache storage type}
\usage{
StorageType
}
\description{
The type of storage to use for the terminology cache.
}
\details{
The following values are supported:
\itemize{
  \item \code{MEMORY} - Use an in-memory cache
  \item \code{DISK} - Use a disk-based cache
}
}
