% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udfs.R
\name{to_array}
\alias{to_array}
\title{Convert a vector to a SQL array literal}
\usage{
to_array(value)
}
\arguments{
\item{value}{A character or numeric vector to be converted}
}
\value{
The \code{quosure} with the SQL array literal that can be used in \code{dplyr::mutate}.
}
\description{
Converts a vector to an expression with the corresponding SQL array literal.
}
