% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udfs.R
\name{tx_subsumes}
\alias{tx_subsumes}
\title{Test subsumption between codings}
\usage{
tx_subsumes(left_codings, right_codings)
}
\arguments{
\item{left_codings}{A Column containing a struct representation of a Coding or an array of Codings.}

\item{right_codings}{A Column containing a struct representation of a Coding or an array of Codings.}
}
\value{
A Column containing the result of the operation (boolean).
}
\description{
Takes two Coding columns as input. Returns a Column that contains a Boolean value,
indicating whether the left Coding subsumes the right Coding.
}
\examples{
\dontrun{
# Test the codings of the Condition `code` for subsumption of a SNOMED CT code.
pc \%>\%
  pathling_example_resource("Condition") \%>\%
  sparklyr::mutate(
    id,
    subsumes = !!tx_subsumes(
      code[["coding"]],
      !!tx_to_snomed_coding("444814009")
    ),
    .keep = "none"
  )
}
}
\seealso{
\href{https://pathling.csiro.au/docs/libraries/terminology#subsumption-testing}{Pathling documentation - Subsumption testing}

Other terminology functions:
\code{\link[=tx_display]{tx_display()}},
\code{\link[=tx_member_of]{tx_member_of()}},
\code{\link[=tx_property_of]{tx_property_of()}},
\code{\link[=tx_subsumed_by]{tx_subsumed_by()}},
\code{\link[=tx_translate]{tx_translate()}}
}
\concept{terminology functions}
