% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir.R
\docType{data}
\name{Version}
\alias{Version}
\title{FHIR versions}
\format{
An object of class \code{list} of length 1.
}
\usage{
Version
}
\description{
The following FHIR versions are supported:
\itemize{
  \item{\code{R4}: FHIR R4}
}
}
\keyword{datasets}
