% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conditions}
\alias{conditions}
\title{Synthetic conditions data}
\format{
An object of class \code{data.frame} with 19 rows and 6 columns.
}
\usage{
conditions
}
\description{
A synthetic data set of simplified and flattened FHIR Condition resources generated by Synthea.
}
\details{
A data frame with 19 rows and 6 columns:
\itemize{
  \item \code{START} - The onset date
  \item \code{STOP} - The abatement date
  \item \code{PATIENT} - The ID of the patient
  \item \code{ENCOUNTER} - The ID of the encounter
  \item \code{CODE} - The SNOMED CT code of the condition
  \item \code{DESCRIPTION} - The display name of the condition
}
}
\keyword{datasets}
