% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{LOINC_URI}
\alias{LOINC_URI}
\title{LOINC system URI}
\format{
An object of class \code{character} of length 1.
}
\usage{
LOINC_URI
}
\description{
The URI of the LOINC code system: \code{http://loinc.org}.
}
\seealso{
\href{https://terminology.hl7.org/LOINC.html}{Using LOINC with HL7 Standards}
}
\keyword{datasets}
