% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir.R
\docType{data}
\name{MimeType}
\alias{MimeType}
\title{FHIR MIME types}
\format{
An object of class \code{list} of length 2.
}
\usage{
MimeType
}
\description{
The following MIME types are supported:
\itemize{
  \item{\code{FHIR_JSON}: FHIR resources encoded as JSON}
  \item{\code{FHIR_XML}: FHIR resources encoded as XML}
}
}
\seealso{
\href{https://hl7.org/fhir/R4/formats.html}{FHIR R4 - Resource Formats}
}
\keyword{datasets}
