% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{pathling_disconnect}
\alias{pathling_disconnect}
\title{Disconnect from the Spark session}
\usage{
pathling_disconnect(pc)
}
\arguments{
\item{pc}{The PathlingContext object.}
}
\value{
No return value, called for side effects only.
}
\description{
Disconnects the Spark connection associated with a Pathling context.
}
\seealso{
Other context lifecycle functions: 
\code{\link{pathling_connect}()},
\code{\link{pathling_disconnect_all}()},
\code{\link{pathling_spark}()}
}
\concept{context lifecycle functions}
