% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{pathling_disconnect_all}
\alias{pathling_disconnect_all}
\title{Disconnect all Spark connections}
\usage{
pathling_disconnect_all()
}
\value{
No return value, called for side effects only.
}
\description{
Disconnect all Spark connections
}
\seealso{
Other context lifecycle functions: 
\code{\link{pathling_connect}()},
\code{\link{pathling_disconnect}()},
\code{\link{pathling_spark}()}
}
\concept{context lifecycle functions}
