% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{pathling_example_resource}
\alias{pathling_example_resource}
\title{Read resource from Pathling example data}
\usage{
pathling_example_resource(pc, resource_name)
}
\arguments{
\item{pc}{The PathlingContext object.}

\item{resource_name}{The name of the resource to read.}
}
\value{
A Spark DataFrame containing the resource data.
}
\description{
Reads a FHIR resource dataframe from the package example data.
}
\details{
The resources are read from the package example data in the \code{extdata/parquet} directory. 
Currently the following resources are available: 'Patient' and 'Condition'.
}
\examples{
\dontshow{if (pathling_is_spark_installed()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pc <- pathling_connect()
pathling_example_resource(pc, 'Condition')
pathling_disconnect(pc)
\dontshow{\}) # examplesIf}
}
\seealso{
Other example functions: 
\code{\link{pathling_examples}()}
}
\concept{example functions}
