% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{pathling_examples}
\alias{pathling_examples}
\title{Get path to Pathling example data}
\usage{
pathling_examples(...)
}
\arguments{
\item{...}{character vector of the path components.}
}
\value{
The path to the examples data.
}
\description{
Construct the path to the package example data in a platform-independent way.
}
\examples{
pathling_examples("ndjson", "Condition.ndjson")
}
\seealso{
Other example functions: 
\code{\link{pathling_example_resource}()}
}
\concept{example functions}
