% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{pathling_is_spark_installed}
\alias{pathling_is_spark_installed}
\title{Check if Spark is installed}
\usage{
pathling_is_spark_installed()
}
\value{
\code{TRUE} if the required version of Spark/Hadoop is installed, \code{FALSE} otherwise.
}
\description{
Checks if the version of Spark/Hadoop required by Pathling is installed.
}
\seealso{
Other installation functions: 
\code{\link{pathling_install_spark}()},
\code{\link{pathling_spark_info}()},
\code{\link{pathling_version}()}
}
\concept{installation functions}
