% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{pathling_spark_info}
\alias{pathling_spark_info}
\title{Get versions of Spark and other dependencies}
\usage{
pathling_spark_info()
}
\value{
A list containing the following keys:
\itemize{
  \item{\code{spark_version}: The version of Spark used by Pathling.}
  \item{\code{scala_version}: The version of Scala used by Pathling.}
  \item{\code{hadoop_version}: The version of Hadoop used by Pathling.}
  \item{\code{hadoop_major_version}: The major version of Hadoop used by Pathling.}
  \item{\code{delta_version}: The version of Delta used by Pathling.}
}
}
\description{
Returns the versions of Spark and Spark packages used by the Pathling R library.
}
\seealso{
Other installation functions: 
\code{\link{pathling_install_spark}()},
\code{\link{pathling_is_spark_installed}()},
\code{\link{pathling_version}()}
}
\concept{installation functions}
