% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{pathling_version}
\alias{pathling_version}
\title{Get version of Pathling}
\usage{
pathling_version()
}
\value{
The version of the Pathling R library.
}
\description{
Get version of Pathling
}
\seealso{
Other installation functions: 
\code{\link{pathling_install_spark}()},
\code{\link{pathling_is_spark_installed}()},
\code{\link{pathling_spark_info}()}
}
\concept{installation functions}
